Replace SystemUsageRegistry with SystemUsageRegistryProvider
